Date: 2/21/15
Created a new merge script called mergeSeeds.py that merges a set of random number seeds. Note that
this script requires a certain directory structure to operate. If the script discovers that the
directory structure is incorrect, it terminates immediately. Also cleaned out old and unecessary
scripts and slightly modified others.